TEP-0100: Embedded TaskRuns and Runs Status in PipelineRuns [Proposal]
In [TEP-0100], we described the challenges caused by embedding the
complete `TaskRun` and `Run` status in `PipelineRuns`: performance
degradation, memory bloat, and lack of extensibility.

In this change we add the proposal to reduce the amount of information
stored about the status of `TaskRuns` and `Runs` to only the names and
the overall status in order to improve performance, reduce memory bloat
and improve extensibility. We also mark the TEP as implementable.

Co-authored-by: Lee Bernick <leebernick@google.com>